requirements.txt
autopep8==1.7.0
click==8.1.3
dnspython==2.3.0
Flask==2.2.2
importlib-metadata==5.0.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
Pillow==9.4.0
pycodestyle==2.9.1
pymongo==4.3.3
pyzbar==0.1.9
RPi.GPIO==0.7.1
toml==0.10.2
Werkzeug==2.2.2
zipp==3.9.0
